+++ modules/apps/layout/layout-set-prototype-web/src/com/liferay/layout/set/prototype/web/portlet/LayoutSetPrototypePortlet.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.layoutsetprototypes;
+package com.liferay.layout.set.prototype.web.portlet;
 
 import com.liferay.portal.NoSuchLayoutSetPrototypeException;
 import com.liferay.portal.RequiredLayoutSetPrototypeException;
